+++ testutils/src/main/java/com/google/android/exoplayer2/testutil/ExoHostedTest.java
@@ -31,7 +31,6 @@ import com.google.android.exoplayer2.analytics.AnalyticsListener;
 import com.google.android.exoplayer2.audio.DefaultAudioSink;
 import com.google.android.exoplayer2.decoder.DecoderCounters;
 import com.google.android.exoplayer2.drm.DrmSessionManager;
-import com.google.android.exoplayer2.drm.FrameworkMediaCrypto;
 import com.google.android.exoplayer2.source.MediaSource;
 import com.google.android.exoplayer2.testutil.HostActivity.HostedTest;
 import com.google.android.exoplayer2.trackselection.DefaultTrackSelector;
@@ -142,7 +141,7 @@ public abstract class ExoHostedTest implements AnalyticsListener, HostedTest {
       pendingSchedule.start(player, trackSelector, surface, actionHandler, /* callback= */ null);
       pendingSchedule = null;
     }
-    DrmSessionManager<FrameworkMediaCrypto> drmSessionManager = buildDrmSessionManager(userAgent);
+    DrmSessionManager drmSessionManager = buildDrmSessionManager(userAgent);
     player.setMediaSource(buildSource(host, Util.getUserAgent(host, userAgent), drmSessionManager));
     player.prepare();
   }
@@ -232,7 +231,7 @@ public abstract class ExoHostedTest implements AnalyticsListener, HostedTest {
     return true;
   }
 
-  protected DrmSessionManager<FrameworkMediaCrypto> buildDrmSessionManager(String userAgent) {
+  protected DrmSessionManager buildDrmSessionManager(String userAgent) {
     // Do nothing. Interested subclasses may override.
     return DrmSessionManager.getDummyDrmSessionManager();
   }
@@ -255,7 +254,7 @@ public abstract class ExoHostedTest implements AnalyticsListener, HostedTest {
   }
 
   protected abstract MediaSource buildSource(
-      HostActivity host, String userAgent, DrmSessionManager<?> drmSessionManager);
+      HostActivity host, String userAgent, DrmSessionManager drmSessionManager);
 
   protected void onPlayerErrorInternal(ExoPlaybackException error) {
     // Do nothing. Interested subclasses may override.